back/requirements.txt
asgiref==3.8.1
bidict==0.23.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
eventlet==0.39.1
greenlet==3.1.1
h11==0.14.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
PyYAML==6.0.2
simple-websocket==1.1.0
sqlparse==0.5.3
uritemplate==4.1.1
wsproto==1.2.0
